Crate tpe

Macros

impl_hash_with_serialized_compressed 🔒

Structs

AggregateDecryptionKey

AggregatePublicKey

CipherText

DecryptionKeyShare

EphemeralPublicKey

EphemeralSignature

G1Affine

This is an element of $\mathbb{G}_1$ represented in the affine coordinate space. It is ideal to keep elements in this representation to reduce memory usage and improve performance through the use of mixed curve model arithmetic.

G2Affine

This is an element of $\mathbb{G}_2$ represented in the affine coordinate space. It is ideal to keep elements in this representation to reduce memory usage and improve performance through the use of mixed curve model arithmetic.

PublicKeyShare

SecretKeyShare

Functions

aggregate_decryption_shares

create_decryption_key_share

decrypt_preimage

derive_agg_decryption_key

derive_ephemeral_sk 🔒

derive_public_key_share

encrypt_preimage

hash_to_message 🔒

lagrange_multipliers 🔒

verify_agg_decryption_key

The function asserts that the ciphertext is valid.

verify_ciphertext

Verifying a ciphertext guarantees that it has not been malleated.

verify_decryption_key_share

The function asserts that the ciphertext is valid.

xor_with_hash 🔒